+++ Cooperators/src/main/java/com/autobon/cooperators/entity/CoopAccount.java
@@ -1,8 +1,15 @@
 package com.autobon.cooperators.entity;
 
 import com.autobon.shared.Crypto;
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
 
 import javax.persistence.*;
+import java.util.ArrayList;
+import java.util.Collection;
 import java.util.Date;
 
 /**
@@ -11,7 +18,10 @@ import java.util.Date;
 
 @Entity
 @Table(name = "t_coop_account")
-public class CoopAccount {
+public class CoopAccount implements UserDetails {
+    private static Logger log = LoggerFactory.getLogger(CoopAccount.class);
+    private static String Token = "Autobon~!@#ABCD=";
+
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column private int id;
@@ -34,6 +44,7 @@ public class CoopAccount {
     @Column
     private boolean gender;
 
+    @JsonIgnore
     @Column
     private String password;
 
@@ -45,6 +56,11 @@ public class CoopAccount {
 
     @Column
     private Date createTime; //注册时间
+
+    @Column
+    private String pushId; // 个推客户端ID, 由手机端更新
+
+    @Column private int statusCode; //状态 0-未审核 1-审核成功 2-审核失败 3-账号禁用
 
     public static String encryptPassword(String password) {
         return Crypto.encryptBySha1(password);
@@ -106,8 +122,67 @@ public class CoopAccount {
         this.gender = gender;
     }
 
+    public static String makeToken(int id) {
+        return "cooperator:" + Crypto.encryptAesBase64(String.valueOf(id), Token);
+    }
+
+    public static int decodeToken(String token) {
+        String[] arr = token.split(":");
+        if (arr.length < 2 || !arr[0].equals("cooperator")) return 0;
+        else token = arr[1];
+        try {
+            return Integer.parseInt(Crypto.decryptAesBase64(token, Token));
+        } catch (Exception ex) {
+            log.info("无效token: " + token);
+        }
+        return 0;
+    }
+
+    @JsonIgnore
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        ArrayList<GrantedAuthority> list = new ArrayList<>();
+        list.add(new GrantedAuthority() {
+            @Override
+            public String getAuthority() {
+                return "COOPERATOR";
+            }
+        });
+        return list;
+    }
+
     public String getPassword() {
         return password;
+    }
+
+    @JsonIgnore
+    @Override
+    public String getUsername() {
+        return phone;
+    }
+
+    @JsonIgnore
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @JsonIgnore
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @JsonIgnore
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @JsonIgnore
+    @Override
+    public boolean isEnabled() {
+        return this.statusCode != 3;
     }
 
     public void setPassword(String password) {
@@ -136,5 +211,21 @@ public class CoopAccount {
 
     public void setCreateTime(Date createTime) {
         this.createTime = createTime;
+    }
+
+    public String getPushId() {
+        return pushId;
+    }
+
+    public void setPushId(String pushId) {
+        this.pushId = pushId;
+    }
+
+    public int getStatusCode() {
+        return statusCode;
+    }
+
+    public void setStatusCode(int statusCode) {
+        this.statusCode = statusCode;
     }
 }
