+++ Cooperators/src/main/resources/db/migration/V20160323.1__t_coop_account.sql
@@ -10,7 +10,10 @@ CREATE TABLE IF NOT EXISTS t_coop_account (
   last_login_time datetime DEFAULT NULL COMMENT '上次登录时间',
   last_login_ip varchar(20) DEFAULT NULL COMMENT '上次登录IP',
   create_time datetime DEFAULT NULL COMMENT '注册时间',
-  PRIMARY KEY (id)
-)DEFAULT CHARSET=utf8 COMMENT='合作商户业务员表';
+  push_id VARCHAR(128) COLLATE utf8_unicode_ci DEFAULT NULL,
+  status_code int(1) DEFAULT NULL COMMENT '状态 0-未审核 1-审核成功 2-审核失败 3-账号禁用',
+  PRIMARY KEY (id),
+  KEY coop_push_id (push_id)
+)DEFAULT CHARSET=utf8 COMMENT='合作商户账户表';
 
-insert  into `t_coop_account`(`id`,`cooperator_id`,`is_main`,`fired`,`phone`,`name`,`gender`,`password`,`last_login_time`,`last_login_ip`,`create_time`) values (1,1,0,0,'13072726003','Tom',1,'7c4a8d09ca3762af61e59520943dc26494f8941b','2015-03-05 11:11:11','192.168.1.1','2015-03-02 11:11:11');
+#insert  into `t_coop_account`(`id`,`cooperator_id`,`is_main`,`fired`,`phone`,`name`,`gender`,`password`,`last_login_time`,`last_login_ip`,`create_time`) values (1,1,0,0,'13072726003','Tom',1,'7c4a8d09ca3762af61e59520943dc26494f8941b','2015-03-05 11:11:11','192.168.1.1','2015-03-02 11:11:11');
