+++ src/main/java/com/autobon/platform/security/TokenAuthenticationProcessingFilter.java
@@ -1,6 +1,8 @@
 package com.autobon.platform.security;
 
+import com.autobon.cooperators.entity.CoopAccount;
 import com.autobon.cooperators.entity.Cooperator;
+import com.autobon.cooperators.service.CoopAccountService;
 import com.autobon.cooperators.service.CooperatorService;
 import com.autobon.staff.entity.Staff;
 import com.autobon.staff.service.StaffService;
@@ -74,8 +76,8 @@ public class TokenAuthenticationProcessingFilter extends AbstractAuthenticationP
                 }
             }
         } else if (token.startsWith("cooperator:")) {
-            int id = Cooperator.decodeToken(token);
-            if (id > 0) user = applicationContext.getBean(CooperatorService.class).get(id);
+            int id = CoopAccount.decodeToken(token);
+            if (id > 0) user = applicationContext.getBean(CoopAccountService.class).getById(id);
         } else if (token.startsWith("staff:")) {
             int id = Staff.decodeToken(token);
             if (id > 0) user = applicationContext.getBean(StaffService.class).get(id);
